refactor: move key handling to separate modules
