+++ src/keymaps/mod.rs
@@ -1,0 +1,3 @@
+pub mod command;
+pub mod normal;
+pub mod search;
